+++ CHANGELOG.md
@@ -6,9 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Add support for [Iceberg connector](https://trino.io/docs/current/connector/iceberg.html) ([#285]).
+- Add support for [Iceberg connector](https://trino.io/docs/current/connector/iceberg.html) ([#286]).
 
-[#285]: https://github.com/stackabletech/trino-operator/pull/285
+[#286]: https://github.com/stackabletech/trino-operator/pull/286
 
 ## [0.6.0] - 2022-09-08