+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Add support for Trino `395-stackable0.1.0` and `396-stackable0.1.0` ([#292]).
 - Add support for [Iceberg connector](https://trino.io/docs/current/connector/iceberg.html) ([#286]).
 - Add support for [TPCH connector](https://trino.io/docs/current/connector/tpch.html) and [TPCDS connector](https://trino.io/docs/current/connector/tpcds.html) ([#293]).
 
@@ -15,6 +16,7 @@ All notable changes to this project will be documented in this file.
 
 [#286]: https://github.com/stackabletech/trino-operator/pull/286
 [#289]: https://github.com/stackabletech/trino-operator/pull/289
+[#292]: https://github.com/stackabletech/trino-operator/pull/292
 [#293]: https://github.com/stackabletech/trino-operator/pull/293
 
 ## [0.6.0] - 2022-09-08